# chore: use a single buffer for field checks