Instead of creating a new string every for each possible version of a field we now (re-)use a single byte array to do the same.

Benchmark comparison is:
```
goos: linux
goarch: amd64
pkg: github.com/terramate-io/tfjson/v2/sanitize
cpu: 13th Gen Intel(R) Core(TM) i7-13620H
                  │   v2.txt    │              v2_pr.txt              │
                  │   sec/op    │   sec/op     vs base                │
LargeChangeset-16   24.55µ ± 6%   13.01µ ± 2%  -46.99% (p=0.000 n=10)

                  │   v2.txt    │             v2_pr.txt              │
                  │    B/op     │    B/op     vs base                │
LargeChangeset-16   7672.0 ± 0%   784.0 ± 0%  -89.78% (p=0.000 n=10)

                  │   v2.txt    │             v2_pr.txt              │
                  │  allocs/op  │ allocs/op   vs base                │
LargeChangeset-16   368.00 ± 0%   42.00 ± 0%  -88.59% (p=0.000 n=10)
```

Before:
```
goos: linux
goarch: amd64
pkg: github.com/terramate-io/tfjson/v2/sanitize
cpu: 13th Gen Intel(R) Core(TM) i7-13620H
BenchmarkLargeChangeset
BenchmarkLargeChangeset-16    	   46780	     22959 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   45748	     23622 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   51043	     24620 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   48404	     23003 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   44822	     25041 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   46743	     24831 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   51252	     25010 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   48674	     23974 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   45248	     24541 ns/op	    7672 B/op	     368 allocs/op
BenchmarkLargeChangeset-16    	   50532	     24560 ns/op	    7672 B/op	     368 allocs/op
PASS
ok  	github.com/terramate-io/tfjson/v2/sanitize	14.217s
```

With PR:
```
goos: linux
goarch: amd64
pkg: github.com/terramate-io/tfjson/v2/sanitize
cpu: 13th Gen Intel(R) Core(TM) i7-13620H
BenchmarkLargeChangeset
BenchmarkLargeChangeset-16    	   88532	     12938 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   88912	     13293 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91579	     13061 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91285	     12980 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91048	     13062 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91712	     12895 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91569	     12948 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91477	     13012 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91344	     13227 ns/op	     784 B/op	      42 allocs/op
BenchmarkLargeChangeset-16    	   91731	     13015 ns/op	     784 B/op	      42 allocs/op
PASS
ok  	github.com/terramate-io/tfjson/v2/sanitize	13.270s
```